Update README.md
This is a website that is still in the works. But it is essentially a portfolio site that is going to have all of my projects that are from this bootcamp and the some that I am currently working on right now. The links that show the name of Pokemon games are placeholders for the projects that will be finished soon.  

This website will be fixed, I've been really busy this past week with work and college but once it is finished you will like it :)

http://127.0.0.1:5500/assets/portfolio.html